Chore: Cookie 관련 메서드 분리
