+++ src/main/java/com/jemyeonso/app/jemyeonsobe/util/CookieUtil.java
@@ -1,0 +1,16 @@
+package com.jemyeonso.app.jemyeonsobe.util;
+
+import jakarta.servlet.http.Cookie;
+import jakarta.servlet.http.HttpServletResponse;
+import org.springframework.stereotype.Component;
+
+@Component
+public class CookieUtil {
+    public void invalidateCookie(HttpServletResponse response, String name) {
+        Cookie cookie = new Cookie(name, null);
+        cookie.setHttpOnly(true);
+        cookie.setPath("/");
+        cookie.setMaxAge(0); // 즉시 만료
+        response.addCookie(cookie);
+    }
+}
